various changes

+++ source/callables/Render.cpp
@@ -12,7 +12,7 @@ void Render::execute()
 		for (unsigned int position = 0; position < m_RectangleShapes->objectCount(); position++)
 		{
 			// check if rectangle is enabled to be rendered and if its on this render layer
-			if (m_RectangleShapes->get(position).isRenderEnabled() && (unsigned int)m_RectangleShapes->get(position).getRenderLayer() == renderLayer)
+			if (m_RectangleShapes->get(position).isRenderEnabled() && m_RectangleShapes->get(position).getRenderLayer() == renderLayer)
 			{
 				// render rectangle shape in this position of the rectangle shape array
 				m_RenderWindow->draw(m_RectangleShapes->get(position));
@@ -23,7 +23,7 @@ void Render::execute()
 		for (unsigned int position = 0; position < m_RectangleShapes->objectCount(); position++)
 		{
 			// check if text is enabled to be rendered and if its on this render layer
-			if (m_Texts->get(position).isRenderEnabled() && (unsigned int)m_Texts->get(position).getRenderLayer() == renderLayer)
+			if (m_Texts->get(position).isRenderEnabled() && m_Texts->get(position).getRenderLayer() == renderLayer)
 			{
 				// render Text in this position of the text array
 				m_RenderWindow->draw(m_Texts->get(position));
